main: extract reload pipeline module
